refactor: icrc canister interaction
