+++ frontend/pages/contacts/components/ICRC/AddContact/index.tsx
@@ -11,7 +11,6 @@ import ContactMainDetails from "./ContactMainDetails";
 import ContactAssetDetails from "./ContactAssetDetails";
 import LoadingLoader from "@components/Loader";
 import { CustomButton } from "@components/Button";
-import { retrieveSubAccountsWithAllowance, retrieveAssetsWithAllowance } from "@/pages/home/helpers/icrc";
 import { AssetContact } from "@redux/models/ContactsModels";
 import { isHexadecimalValid, validateSubaccounts } from "@/utils/checkers";
 import clsx from "clsx";
@@ -19,6 +18,7 @@ import { validatePrincipal } from "@/utils/identity";
 import { SupportedStandardEnum } from "@/@types/icrc";
 import { db } from "@/database/db";
 import { getAccountIdentifier } from "@/utils";
+import { retrieveAssetsWithAllowance, retrieveSubAccountsWithAllowance } from "@pages/home/helpers/icrc/";
 
 interface AddContactProps {
   setAddOpen(value: boolean): void;
