+++ frontend/pages/home/components/ICRC/allowance/CreateForm/index.tsx
@@ -8,7 +8,7 @@ import ExpirationFormItem from "./ExpirationFormItem";
 import { AllowanceValidationErrorsEnum } from "@/@types/allowance";
 import { useTranslation } from "react-i18next";
 import { CustomButton } from "@components/Button";
-import { getAllowanceDetails } from "@pages/home/helpers/icrc";
+import { getAllowanceDetails } from "@pages/home/helpers/icrc/";
 import { validatePrincipal } from "@/utils/identity";
 import { isHexadecimalValid } from "@/utils/checkers";
 import {
