+++ frontend/pages/home/hooks/useSend.ts
@@ -1,7 +1,7 @@
 import { toFullDecimal } from "@/utils";
 import { useAppSelector } from "@redux/Store";
 import { useMemo } from "react";
-import { getAllowanceDetails } from "../helpers/icrc";
+import { getAllowanceDetails } from "../helpers/icrc/";
 import { TransactionSenderOptionEnum } from "@/@types/transactions";
 import { validatePrincipal } from "@/utils/identity";
 import { isHexadecimalValid } from "@/utils/checkers";
@@ -48,6 +48,8 @@ export default function useSend() {
   const getTransactionFee = () => {
     const hasTransactionFee = updateSubAccount?.transaction_fee !== undefined;
     const decimalPlaces = updateAsset?.decimal ?? 0;
+
+    // TODO: if not call the getTransactionFeeFromLedger
 
     return hasTransactionFee
       ? toFullDecimal(updateSubAccount.transaction_fee, Number(decimalPlaces))
